Make --owner optional for fetch (baby step)
This is a very gentle step towards relaxing the requirements on the
options passed to the fetch (and in the future, the sync) command on the cli.
This step makes the --owner option optional and compensates for an
absent --owner with a default value.

TO-DO: Read the owner from the current working (git) directory.